requirements.txt
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.17.1
azure-mgmt-compute==33.0.0
azure-mgmt-core==1.4.0
azure-mgmt-network==26.0.0
azure-mgmt-resource==23.1.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
idna==3.10
isodate==0.6.1
msal==1.31.0
msal-extensions==1.2.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3

azure-identitypandas==2.2.2
azure-storage-blob==12.23.0
matplotlib==3.9.2
